So quieting the quoted warnings results in compile errors if using the static libs to build lsar/unar. Got it.

+++ CSJSONPrinter.h
@@ -63,6 +63,6 @@
 -(void)printSeparatorIfNeeded;
 
 -(NSString *)stringByEscapingString:(NSString *)string;
--(NSString *)stringByEncodingBytes:(const uint8_t *)bytes length:(int)length;
+-(NSString *)stringByEncodingBytes:(const uint8_t *)bytes length:(NSInteger)length;
 
 @end
